Update submodule to latest commit, identify meta.txt/data.csv files
I fixed the code to update the submodule ipcc-fact-checking:
the previous code was meant to update the submodule in reaction
to a change of the upstream parent repository ipcc-facts-import.
Instead, updating the submodule required, from within the submodule
directory, to ensure that the master branch was selected and to pull
(fetch and merge) the latest changes into the master branch.

What is not included is the update of the parent repository
ipcc-facts-import to move the submodule reference to the latest
commit. This requires a commit in the parent module, which should
not be automated, but done manually by a human.

I also added code to locate meta.txt and data.csv within datasets
folder in ipcc-fact-checking, and to extract the source, document
and dataset identifiers (folder names) from the files's parent path.