Report missing end of fields and skip end of line
I encountered an error when a newline character was left in a field;
the value was quoted but the newline character was interpreted as the
end of the record. I do not expect to handle this case differently;
the error reporting should help to locate the source of the failure
and fix it in the data.csv file of origin.